+++ lib/constant.d.ts
@@ -1,8 +1,0 @@
-export declare const DIALOG_FLOW_ACTION = "__dfAction__";
-export declare const DIALOG_FLOW_INTENT = "__dfIntent__";
-export declare const DIALOG_FLOW_PARAMS = "__dfParams__";
-export declare const nestMetadata: {
-    COMPONENTS: string;
-};
-export declare const PATH_METADATA = "path";
-export declare const METHOD_METADATA = "method";