+++ lib/decorators/dialog-flow-action.decorator.d.ts
@@ -1,2 +1,0 @@
-import 'reflect-metadata';
-export declare const DialogFlowAction: (action: string) => (target: Object, key: string | symbol, descriptor: PropertyDescriptor) => PropertyDescriptor;