+++ lib/decorators/dialog-flow-param.decorator.d.ts
@@ -1,2 +1,0 @@
-import 'reflect-metadata';
-export declare const DialogFlowParam: (property?: string) => (target: any, key: any, index: any) => any;