+++ lib/middlewares/dialog-flow-authorization.middleware.d.ts
@@ -1,4 +1,0 @@
-import { NestMiddleware } from '@nestjs/common';
-export declare class DialogFlowAuthorizationMiddleware implements NestMiddleware {
-    resolve(): (req: any, res: any, next: any) => any;
-}