+++ lib/utils.d.ts
@@ -1,1 +1,0 @@
-export declare function applyParamsMetadataDecorator(paramsMetadata: any[], args: any[]): any[];