+++ src/core/index.ts
@@ -1,2 +1,1 @@
-export * from './scanner';
 export * from './handlers';